third_party/blink/web_tests/external/wpt/css/css-color/webkit-focus-ring-color.tentative.html

<!doctype html>
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1718934">
<title>-webkit-focus-ring-color is supported for compat</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(CSS.supports("color", "-webkit-focus-ring-color"));
});
</script>
